refactor(knowledge): split the code wiki endpoints and repository access out
`knowledge.py` was 2471 lines and `wiki_service.py` 1389, both past the
1000-line rule, and both had a seam worth cutting along rather than an
arbitrary line count to get under.

Code wiki endpoints move to `knowledge_code_wiki.py`. They are about
repositories and runs -- resolving one, claiming it, starting a run -- while
the module they leave is about documents. Registration order is now
load-bearing rather than incidental: `/code-wikis` has to be matched before
`/{knowledge_base_id}`, which would otherwise read "code-wikis" as an id, so
there is a test asserting it.

Repository access moves to `wiki_repository_access.py`. It answers one
question -- may this user read the repository behind this project -- and
answers it for every read path, so it must not have two answers. It talks to
git providers, caches and tokens and knows nothing about pages, which is a
different subject from the service that stores them.

Behaviour is unchanged. One docstring written during the move claimed a
cache miss falls back to the provider; it does not, and the corrected text
says what the code does and why.
